code/jockey-server/requirements.txt
jupyterlab>3.0
ffmpeg-python==0.2.0
grequests==0.7.0
tabulate==0.9.0
python-dotenv==1.0.1
rich==13.7.0
langchain-openai==0.1.22
langchain==0.2.14
langgraph>=0.2.32,<0.3.0
langgraph-cli
langgraph-sdk
openai==1.55.3

